Add files via upload
The corpus contains running text of 5,042 words of Urdu, in which all ambiguous words (856 instances) are manually tagged with senses from the Urdu Lughat dictionary. A range of baseline WSD models based on n-grams are applied to the corpus and the best performance (with Accuracy of 57.71%) is achieved using word 4-grams.